Extract TimeInterval as unique function component from home page
