+++ src/pages/home.tsx
@@ -8,6 +8,7 @@ import ReactLoading from "react-loading";
 import { ArticleType } from "../utils/articleType";
 import SavedNewsBtn from "../components/savedNewsBtn";
 import timestampConvertDate from "../utils/timeStampConverter";
+import TimeInterval from "../components/timeInterval";
 import CategoryTag from "../components/categoryTag";
 import ViewCount from "../components/viewCountDiv";
 import gainViews from "../utils/gainViews";
@@ -336,44 +337,44 @@ const ScrollTarget = styled.div`
   background-color: #f35b03;
 `;
 
-const TimeTag = styled.div`
-  /* width: 200px; */
-  padding-left: 5px;
-  position: absolute;
-  text-align: center;
-  bottom: -33px;
-  left: 0px;
-  z-index: 4;
-  font-size: 16px;
-  line-height: 20px;
-  border-left: 2px solid #000000;
-  font-family: "Quicksand", sans-serif;
-  font-weight: 500;
-  @media screen and (max-width: 1280px) {
-    bottom: -18px;
-    font-size: 10px;
-    line-height: 12px;
-  }
-`;
+// const TimeTag = styled.div`
+//   /* width: 200px; */
+//   padding-left: 5px;
+//   position: absolute;
+//   text-align: center;
+//   bottom: -33px;
+//   left: 0px;
+//   z-index: 4;
+//   font-size: 16px;
+//   line-height: 20px;
+//   border-left: 2px solid #000000;
+//   font-family: "Quicksand", sans-serif;
+//   font-weight: 500;
+//   @media screen and (max-width: 1280px) {
+//     bottom: -18px;
+//     font-size: 10px;
+//     line-height: 12px;
+//   }
+// `;
 
-const TimeTagEven = styled.div`
-  position: absolute;
-  padding-left: 5px;
-  text-align: center;
-  top: -33px;
-  left: 0px;
-  z-index: 4;
-  font-size: 16px;
-  line-height: 20px;
-  border-left: 2px solid #000000;
-  font-family: "Quicksand", sans-serif;
-  font-weight: 500;
-  @media screen and (max-width: 1280px) {
-    top: -18px;
-    font-size: 10px;
-    line-height: 12px;
-  }
-`;
+// const TimeTagEven = styled.div`
+//   position: absolute;
+//   padding-left: 5px;
+//   text-align: center;
+//   top: -33px;
+//   left: 0px;
+//   z-index: 4;
+//   font-size: 16px;
+//   line-height: 20px;
+//   border-left: 2px solid #000000;
+//   font-family: "Quicksand", sans-serif;
+//   font-weight: 500;
+//   @media screen and (max-width: 1280px) {
+//     top: -18px;
+//     font-size: 10px;
+//     line-height: 12px;
+//   }
+// `;
 
 const FlyBackBtn = styled.div`
   width: 50px;
@@ -848,30 +849,6 @@ function Home() {
     setArticles(updatedArticles);
   }
 
-  function timeInterval(time: number, index: number) {
-    const interval = (Date.now() - time) / 1000;
-    const hours = Math.floor(interval / 3600);
-    if (hours < 24) {
-      return index % 2 === 0 ? (
-        <TimeTag>{hours}小時前</TimeTag>
-      ) : (
-        <TimeTagEven>{hours}小時前</TimeTagEven>
-      );
-    } else if (hours <= 48) {
-      return index % 2 === 0 ? (
-        <TimeTag>{Math.round(hours / 24)}天前</TimeTag>
-      ) : (
-        <TimeTagEven>{Math.round(hours / 24)}天前</TimeTagEven>
-      );
-    } else {
-      return index % 2 === 0 ? (
-        <TimeTag>{timeExpression(time)}</TimeTag>
-      ) : (
-        <TimeTagEven>{timeExpression(time)}</TimeTagEven>
-      );
-    }
-  }
-
   function cardOnLoad() {
     return Array.from({
       length: 12,
@@ -1099,7 +1076,7 @@ function Home() {
                                   />
                                 </SavedNewsDiv>
                               </UserInteractDiv>
-                              {timeInterval(article.publishedAt, index)}
+                              {TimeInterval(article.publishedAt, index)}
 
                               {index % 2 === 0 ? (
                                 <SourceTag>{article.source["name"]}</SourceTag>
